+++ geode-core/src/main/java/org/apache/geode/distributed/internal/ConnectionConfig.java
@@ -14,7 +14,7 @@
  */
 package org.apache.geode.distributed.internal;
 
-import org.apache.geode.distributed.internal.membership.QuorumChecker;
+import org.apache.geode.distributed.internal.membership.gms.api.QuorumChecker;
 
 /**
  * Contains the distribution config and internal properties for connecting.
