+++ geode-core/src/test/java/org/apache/geode/distributed/internal/ConnectionConfigImplTest.java
@@ -24,7 +24,7 @@ import java.util.Properties;
 
 import org.junit.Test;
 
-import org.apache.geode.distributed.internal.membership.QuorumChecker;
+import org.apache.geode.distributed.internal.membership.gms.api.QuorumChecker;
 
 /**
  * Unit tests for {@link ConnectionConfigImpl}.
